+++ docs/en/docs/release-notes.md
@@ -21,6 +21,7 @@ hide:
 - Added `python-multipart` as part of the `all` and `standard` Lilya packages.
 - Replaced `python-multipart` with a fully native multipart, urlencoded, and octet-stream form parser.
 - Improved RFC 5987 parameter decoding for proper handling of UTF-8 filenames and headers.
+- `AuthenticationError` exception is now located in `lilya.exceptions`.
 
 
 ## 0.20.5